backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.24.0
pydantic==2.8.0
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.2.2
python-multipart==0.0.8
requests==2.31.0